git-svn-id: http://kos.informatik.uni-osnabrueck.de/software/meshing@2391 bea721bf-f9f1-0310-a2d5-e0bfdc8dd81e

+++ src/lib3d/TriangleMesh.cpp
@@ -48,7 +48,7 @@ void TriangleMesh::finalize(){
 		colors[3 * i + 2] = 0.0f;
 	}
 
-	int iii;
+	size_t iii;
 
 	for(size_t i = 0; i < triangle_buffer.size(); i++){
 		iii = 3 * i;
